test_requirements.txt
 
# uvicorn==0.8.3
# fastapi==0.31.0
requests
pytest
tenacity
python-dotenv==0.10.1
pony==0.7.10
bcrypt==3.1.7
passlib==1.7.1
pyjwt==1.7.1
python-multipart==0.0.5


